Create POC_ManageEngine.py
This is a proof-of-concept Python script that exploits a SQL Injection vulnerability in the AMUserResourcesSyncServlet endpoint of ManageEngine Applications Manager (versions 11.0 to <14.0). The exploit leverages unauthenticated SQL injection to define a User-Defined Function (UDF), which loads a malicious DLL over SMB to achieve remote code execution (RCE) on the underlying Windows system.